pyproject.toml
[project]
name = "fiap-tech-challenge-04"
version = "1.0.0"
description = ""
authors = [
    {name = "natokratos",email = "darktimeangel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "boto3 (>=1.37.22,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "endpoints (>=7.3.1,<8.0.0)",
    "setuptools (>=78.1.0,<79.0.0)",
    "standard-imghdr (>=3.13.0,<4.0.0)",
    "selenium (>=4.30.0)",
    "pandas (>=2.2.3)",
    "psycopg2-binary (>=2.9.10)",
    "scikit-learn (>=1.6)",
    "matplotlib (>=3.10.1)"
]

[tool.poetry]
packages = [
    { include = "src/main.py" },
    { include = "src/api.py" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
